+++ src/main/java/wooteco/subway/controller/ExceptionAdvice.java
@@ -1,8 +1,9 @@
-package wooteco.subway.exception;
+package wooteco.subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import wooteco.subway.exception.WebException;
 
 @RestControllerAdvice
 public class ExceptionAdvice {
