+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -1,8 +1,11 @@
-package wooteco.subway.station;
+package wooteco.subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.dto.request.StationRequest;
+import wooteco.subway.dto.response.StationResponse;
+import wooteco.subway.service.StationService;
 
 import java.net.URI;
 import java.util.List;
