+++ src/main/java/wooteco/subway/domain/station/Station.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.domain.station;
 
 public class Station {
     private Long id;
