+++ src/main/java/wooteco/subway/service/LineService.java
@@ -1,7 +1,11 @@
-package wooteco.subway.line;
+package wooteco.subway.service;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import wooteco.subway.dao.LineDao;
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.line.LineName;
+import wooteco.subway.dto.response.LineResponse;
 import wooteco.subway.exception.line.LineDuplicateException;
 import wooteco.subway.exception.line.LineNotExistException;
