+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,7 +1,11 @@
-package wooteco.subway.station;
+package wooteco.subway.service;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import wooteco.subway.dao.StationDao;
+import wooteco.subway.domain.station.Station;
+import wooteco.subway.domain.station.StationName;
+import wooteco.subway.dto.response.StationResponse;
 import wooteco.subway.exception.station.StationDuplicateException;
 
 import java.util.List;
