--- src/main/java/wooteco/subway/station/StationDao.java
@@ -1,50 +1,0 @@
-package wooteco.subway.station;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.util.List;
-import java.util.Objects;
-
-@Repository
-public class StationDao {
-    private final JdbcTemplate jdbcTemplate;
-
-    public StationDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public Station insert(StationName name) {
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-        String query = "INSERT INTO station (name) VALUES (?)";
-        jdbcTemplate.update((Connection con) -> {
-            PreparedStatement pstmt = con.prepareStatement(
-                    query,
-                    new String[]{"id"});
-            pstmt.setString(1, name.getName());
-            return pstmt;
-        }, keyHolder);
-
-        final long id = Objects.requireNonNull(keyHolder.getKey()).longValue();
-        return new Station(id, name);
-    }
-
-    public List<Station> findAll() {
-        String query = "SELECT * FROM station";
-        return jdbcTemplate.query(query,
-                (resultSet, rowNum) -> new Station(
-                        resultSet.getLong("id"),
-                        resultSet.getString("name")
-                )
-        );
-    }
-
-    public void deleteById(Long id) {
-        String query = "DELETE FROM station WHERE id = ?";
-        jdbcTemplate.update(query, id);
-    }
-}
