+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.dto.response.LineResponse;
 
 import java.util.Arrays;
 import java.util.HashMap;
