+++ src/test/java/wooteco/subway/line/LineDaoTest.java
@@ -8,6 +8,9 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.Rollback;
+import wooteco.subway.dao.LineDao;
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.line.LineName;
 
 import java.util.List;
