+++ src/test/java/wooteco/subway/station/StationNameTest.java
@@ -2,6 +2,7 @@ package wooteco.subway.station;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.station.StationName;
 import wooteco.subway.exception.station.StationNamePatternException;
 
 import static org.assertj.core.api.Assertions.assertThat;
